Add transaction options
This update extends the transaction value to include optional options for defining a customer namer and a custom query log function. This enables the user to customize how queries are named and to log the query results to aid with debugging or general information.

An Open function, mirroring sql.Open, is added as a convenience to add usable options to a query transaction.